requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
selenium==4.16.0

Flask==3.0.0
psycopg2-binary==2.9.9
# Add other dependencies as needed
